deleted MetricUtils, called dotPlot from runProali
